eve7: use openui5 loader for all scripts
Need to set correct dependency for JSROOT modules as well
