+++ ui5/eve7/index.html
@@ -5,8 +5,6 @@
       <meta http-equiv="X-UA-Compatible" content="IE=edge">
       <title>Eve7</title>
       <script type="text/javascript" src="jsrootsys/scripts/JSRootCore.js"></script>
-      <!-- script type="text/javascript" src="evedir/Colors.js"></script-->
-      <!-- script type="text/javascript" src="evedir/EveManager.js"></script-->
       <!-- link rel="stylesheet" type="text/css" href="evedir/eve.css"/-->
       <style>
          html { height: 100%; }
@@ -30,13 +28,7 @@
 
       <script type='text/javascript'>
 
-        // var evedir = 
-        
-        var evedir = JSROOT.source_dir.replace(/jsrootsys/g, "rootui5sys/eve7");
-
         function InitUI(handle) {
-
-           // jQuery.sap.registerModulePath("eve", evedir);
 
            var kind = JSROOT.GetUrlOption('view');
 
@@ -59,7 +51,7 @@
         if (JSROOT.GetUrlOption('libs')!==null) JSROOT.use_full_libs = true;
 
         JSROOT.ConnectWebWindow({
-           prereq: "openui5;user:" + evedir + "lib/EveManager.js",
+           prereq: "openui5",
 //           openui5src: "jsroot",    // use JSROOT provided package, experimental
            openui5src: "https://openui5.hana.ondemand.com/",
            openui5libs: "sap.m, sap.ui.layout, sap.ui.table, sap.ui.commons", // customize openui5 libs
